feat: various refactorings
